requirements.txt
Django==3.2.12
django_haystack==3.1.1
zeep==4.1.0
